+++ Kernel/GenericInterface/Operation/Ticket/TicketUpdate.pm
@@ -12,7 +12,7 @@ package Kernel::GenericInterface::Operation::Ticket::TicketUpdate;
 use strict;
 use warnings;
 
-use Kernel::System::VariableCheck qw(IsArrayRefWithData IsHashRefWithData IsStringWithData);
+use Kernel::System::VariableCheck qw(IsArrayRefWithData IsHashRefWithData IsString);
 
 use base qw(
     Kernel::GenericInterface::Operation::Common
@@ -1039,10 +1039,14 @@ sub _CheckDynamicField {
 
     # check DynamicField item internally
     for my $Needed (qw(Name Value)) {
-        if ( !defined $DynamicField->{$Needed} || !IsStringWithData( $DynamicField->{$Needed} ) ) {
+        if (
+            !defined $DynamicField->{$Needed}
+            || ( !IsString( $DynamicField->{$Needed} ) && ref $DynamicField->{$Needed} ne 'ARRAY' )
+            )
+        {
             return {
                 ErrorCode    => 'TicketUpdate.MissingParameter',
-                ErrorMessage => "TicketUpdate: DynamicField->$Needed  parameter is missing!",
+                ErrorMessage => "TicketUpdate: DynamicField->$Needed parameter is missing!",
             };
         }
     }
@@ -1124,7 +1128,7 @@ sub _CheckAttachment {
         if ( !$Attachment->{$Needed} ) {
             return {
                 ErrorCode    => 'TicketUpdate.MissingParameter',
-                ErrorMessage => "TicketUpdate: Attachment->$Needed  parameter is missing!",
+                ErrorMessage => "TicketUpdate: Attachment->$Needed parameter is missing!",
             };
         }
     }